requirements.txt
openmeteo_requests
requests_cache
retry_requests
pandas
matplotlib
seaborn